feat: Dropped curbs and valid labels only processed
Added a new function in called 'valid_labels()'. This function takes all the pano ids in the dataframe (set) and then for each id checks if a corresponding image exits. This information is used to create a list of all valid pano ids for images (existences). Returns as list. 

In the 'main()' function a dataframe filter has been added so that only 'dropped curb' labels (==1) are kept.